chore: Remove crtp from file names
