+++ test/prelude_test.hpp
@@ -3,26 +3,19 @@
 
 #include "catch2/catch_test_macros.hpp"
 
-#include "efp.hpp"
-#include "test_common.hpp"
+#include "prelude.hpp"
 
 using namespace efp;
-
-// ! WIP
 
 TEST_CASE("compose")
 {
     SECTION("pure0")
     {
         auto add_1 = [](int x)
-        {
-            return x + 1;
-        };
+        { return x + 1; };
 
         auto times_2 = [](int x)
-        {
-            return x * 2;
-        };
+        { return x * 2; };
 
         auto composed = compose(add_1, times_2);
         // auto composed_with_operator = add_1 * times_2;
@@ -34,14 +27,10 @@ TEST_CASE("compose")
     SECTION("pure1")
     {
         auto add_1 = [](int x)
-        {
-            return x + 1;
-        };
+        { return x + 1; };
 
         auto devide_2 = [](int x)
-        {
-            return x / 2.;
-        };
+        { return x / 2.; };
 
         auto composed = compose(devide_2, add_1);
         CHECK(composed(2) == 1.5);
@@ -54,269 +43,123 @@ TEST_CASE("compose")
         auto i_plusplus = [&](int x)
         {
             i += x;
-            return i;
-        };
+            return i; };
 
         auto times_2 = [](int x)
-        {
-            return x * 2;
-        };
+        { return x * 2; };
 
         auto composed = compose(times_2, i_plusplus);
         CHECK(composed(2) == 4);
         CHECK(i == 2);
     }
 
-    // SECTION("n-ary")
-    // {
-    //     auto add_1 = [](int x)
-    //     {
-    //         return x + 1;
-    //     };
+    SECTION("n-ary")
+    {
+        auto add_1 = [](int x)
+        { return x + 1; };
 
-    //     auto devide_2 = [](int x)
-    //     {
-    //         return x / 2.;
-    //     };
+        auto devide_2 = [](int x)
+        { return x / 2.; };
 
-    //     auto composed = compose(devide_2, add_1, square<int>);
-    //     CHECK(composed(2) == 2.5);
-    // }
+        auto square = [](int x)
+        { return x * x; };
+
+        const auto composed = compose(devide_2, add_1, square);
+        CHECK(composed(2) == 2.5);
+    }
 }
-
-// TEST_CASE("IsStaticCapacity")
-// {
-//     CHECK(IsStaticCapacity<int[3]>::value == true);
-//     CHECK(IsStaticCapacity<std::array<int, 3>>::value == true);
-//     CHECK(IsStaticCapacity<std::vector<int>>::value == false);
-// }
-
-// TEST_CASE("AreAllStaticCapacity")
-// {
-//     CHECK(AreAllStaticCapacity<int[3]>::value == true);
-//     CHECK(AreAllStaticCapacity<std::array<int, 3>>::value == true);
-//     CHECK(AreAllStaticCapacity<int[3], std::array<int, 3>>::value == true);
-//     CHECK(AreAllStaticCapacity<std::vector<int>>::value == false);
-//     CHECK(AreAllStaticCapacity<std::array<int, 3>, std::vector<int>>::value == false);
-// }
 
 TEST_CASE("append")
 {
     SECTION("Array")
     {
         Array<double, 6> res{1., 2., 3., 1., 2., 3.};
-        CHECK(append(std_array_3, array_3) == res);
+        CHECK(append(array_3, array_3) == res);
     }
 
     SECTION("ArrVec")
     {
         ArrVec<double, 6> res{1., 2., 3., 1., 2., 3.};
-        CHECK(append(array_vector_3, array_3) == res);
+        CHECK(append(arrvec_3, array_3) == res);
     }
 
     SECTION("Array")
     {
         Vector<double> res{1., 2., 3., 1., 2., 3.};
-        CHECK(append(std_array_3, vector_3) == res);
+        CHECK(append(array_3, vector_3) == res);
     }
 }
 
 TEST_CASE("for_each")
 {
+    Array<int, 3> crtp_array{1, 2, 3};
+
     int res_1 = 0;
-    auto add_1 = [&](int x)
-    {
-        res_1 += x;
-    };
+    auto add_to_res_1 = [&](int x)
+    { res_1 += x; };
 
-    for_each(add_1, c_array_3);
+    for_each(add_to_res_1, crtp_array);
     CHECK(res_1 == 6);
-
-    double res_2 = 0.;
-    auto add_2 = [&](double x)
-    {
-        res_2 += x;
-    };
-
-    for_each(add_2, std_array_3);
-    CHECK(res_2 == 6.);
-
-    double res_3 = 0.;
-    auto add_3 = [&](double x)
-    {
-        res_3 += x;
-    };
-
-    for_each(add_3, std_vector_3);
-    CHECK(res_3 == 6.);
-
-    double res_4 = 0.;
-    auto add_product = [&](double x1, double x2)
-    {
-        res_4 += x1 * x2;
-    };
-
-    for_each(add_product, std_vector_3, c_array_3);
-    CHECK(res_4 == 14.);
-
-    double res_5 = 0.;
-
-    auto add_product_2 = [&](double x1, double x2)
-    {
-        res_5 += x1 * x2;
-    };
-
-    for_each(add_product_2, vector_view_3, vector_view_5);
-    CHECK(res_5 == 14.);
-}
-
-TEST_CASE("for_eachi")
-{
-    int res_1[3];
-    auto set_42 = [](int &x)
-    {
-        x = 42;
-    };
-
-    for_eachi(set_42, res_1);
-
-    const auto is_42 = [](int x)
-    { return x == 42; };
-
-    CHECK(foldl(and_v, true, map(is_42, res_1)) == true);
-
-    std::array<int, 3> res_2;
-
-    for_eachi(set_42, res_2);
-    CHECK(res_2 == std::array<int, 3>{42, 42, 42});
-
-    Vector<int> res_3{0, 0, 0};
-
-    for_eachi(set_42, res_3);
-    CHECK(res_3 == Vector<int>{42, 42, 42});
-
-    SECTION("ArrayView")
-    {
-        int c_arr[3] = {0, 1, 2};
-        ArrayView<int, 3> arr_view{&c_arr[0]};
-
-        for_eachi([](int &x)
-                  { x = 2 * x; },
-                  arr_view);
-
-        CHECK((c_arr[0] == 0 && c_arr[1] == 2 && c_arr[2] == 4));
-    }
 }
 
 TEST_CASE("map")
 {
-    SECTION("c style array")
-    {
-        auto square = [](int x)
-        {
-            return x * x;
-        };
+    Array<int, 3> crtp_array{1, 2, 3};
 
-        CHECK(map(square, c_array_3) == Array<int, 3>{1, 4, 9});
-    }
+    auto times_2 = [](int x)
+    { return 2 * x; };
 
-    SECTION("std::vector")
-    {
-        auto square = [](double x)
-        {
-            return x * x;
-        };
-
-        CHECK(map(square, std_vector_3) == Vector<double>{1., 4., 9.});
-    }
-
-    SECTION("binary of std::array, std::vecor")
-    {
-        auto plus = [](double lhs, double rhs)
-        {
-            return lhs + rhs;
-        };
-
-        CHECK(map(plus, std_array_3, std_vector_3) == Vector<double>{2., 4., 6.});
-    }
+    CHECK(map(times_2, crtp_array) == Array<int, 3>{2, 4, 6});
 }
 
 TEST_CASE("filter")
 {
-    SECTION("even c style array")
+    Array<int, 3> crtp_array{1, 2, 3};
+
+    auto is_even = [](int x)
     {
-        auto is_even = [](double x)
-        {
-            return (int)x % 2 == 0;
-        };
+        return x % 2 == 0 ? true : false;
+    };
 
-        CHECK(filter(is_even, c_array_3) == ArrVec<double, 3>{2.});
-    }
+    ArrVec<int, 3> ref{};
+    ref.resize(1);
+    ref[0] = 2;
 
-    SECTION("odd std::vector")
-    {
-        auto is_odd = [](double x)
-        {
-            return (int)x % 2 != 0;
-        };
-
-        CHECK(filter(is_odd, std_vector_3) == Vector<double>{1., 3.});
-    }
+    CHECK(filter(is_even, crtp_array) == ref);
 }
 
 TEST_CASE("foldl")
 {
-    SECTION("plus on c style arary")
-    {
-        CHECK(foldl(plus, 0, c_array_5) == 15);
-    }
+    ArrVec<int, 5> crtp_arrayvec{1, 2, 3, 4};
+    auto plus = [](int lhs, int rhs)
+    { return lhs + rhs; };
 
-    SECTION("times on std::vectors arary")
-    {
-        CHECK(foldl(times, 1., std_vector_3) == 6.);
-    }
+    CHECK(foldl(plus, 0, crtp_arrayvec) == 10);
 }
 
 TEST_CASE("foldr")
 {
-    SECTION("plus on c style arary")
-    {
-        auto plus = [](int lhs, int rhs)
-        {
-            return lhs + rhs;
-        };
-        CHECK(foldr(plus, 0, c_array_5) == 15);
-    }
+    ArrVec<int, 5> crtp_arrayvec{1, 2, 3, 4};
+    auto plus = [](int lhs, int rhs)
+    { return lhs + rhs; };
 
-    SECTION("times on std::vectors arary")
-    {
-        auto times = [](double lhs, double rhs)
-        {
-            return lhs * rhs;
-        };
-        CHECK(foldr(times, 1., std_vector_3) == 6.);
-    }
+    CHECK(foldr(plus, 0, crtp_arrayvec) == 10);
 }
 
 TEST_CASE("from_function")
 {
+    auto plus_one = [](int i)
+    {
+        return i + 1;
+    };
+
     SECTION("Array")
     {
-        auto plus_one = [](int i)
-        {
-            return i + 1;
-        };
-
         CHECK(from_function(IntegralConst<int, 3>{}, plus_one) == Array<int, 3>{1, 2, 3});
     }
 
-    SECTION("DynamicArray")
+    SECTION("Vector")
     {
-        auto plus_one = [](int i)
-        {
-            return i + 1;
-        };
-
         CHECK(from_function(3, plus_one) == Vector<int>{1, 2, 3});
     }
 }
@@ -347,64 +190,66 @@ TEST_CASE("for_each_with_index")
         res[i] = x0 * x1;
     };
 
-    for_each_with_index(product, c_array_3, std_vector_5);
+    for_each_with_index(product, array_3, vector_5);
     CHECK(res == Array<int, 3>{1, 4, 9});
 }
 
 TEST_CASE("cartesian_for_each")
 {
-    std::vector<int> res;
-    auto as = std::array<int, 2>{1, 2};
-    auto bs = std::array<int, 2>{1, 3};
+    Vector<int> res{};
+    auto as = Array<int, 2>{1, 2};
+    auto bs = Array<int, 2>{1, 3};
 
     auto push_res = [&](int x0, int x1)
     {
-        res.push_back(x0 * x1);
+        res.push_back(10 * x0 + x1);
     };
 
     cartesian_for_each(push_res, as, bs);
-    CHECK(res == std::vector<int>{1, 3, 2, 6});
+    CHECK(res == Vector<int>{11, 13, 21, 23});
 }
 
 TEST_CASE("map_with_index")
 {
-    SECTION("c style array")
+    SECTION("array")
     {
         auto times = [](int lhs, int rhs)
         {
             return lhs * rhs;
         };
 
-        CHECK(map_with_index(times, c_array_3) == Array<int, 3>{0, 2, 6});
+        CHECK(map_with_index(times, array_3) == Array<int, 3>{0, 2, 6});
     }
 
-    SECTION("std::vector")
+    SECTION("vector")
     {
         auto times = [](double lhs, double rhs)
         {
             return lhs * rhs;
         };
 
-        CHECK(map_with_index(times, std_vector_3) == Vector<double>{0., 2., 6.});
+        CHECK(map_with_index(times, vector_3) == Vector<double>{0., 2., 6.});
     }
 
-    SECTION("binary of std::array, std::vecor")
+    SECTION("binary hetero")
     {
         auto product = [](int i, double a, double b)
         {
             return i * a * b;
         };
 
-        CHECK(map_with_index(product, std_array_3, std_vector_3) == Vector<double>{0., 4., 18.});
+        CHECK(map_with_index(product, array_3, vector_3) == Vector<double>{0., 4., 18.});
     }
 }
 
 TEST_CASE("cartesian_map")
 {
-    auto as = std::array<int, 2>{1, 2};
-    auto bs = std::array<int, 2>{1, 3};
+    auto as = Array<int, 2>{1, 2};
+    auto bs = Array<int, 2>{1, 3};
 
-    auto res = cartesian_map(times<int, int>, as, bs);
+    auto res = cartesian_map([](int a, int b)
+                             { return a * b; },
+                             as, bs);
     CHECK(res == Array<int, 4>{1, 3, 2, 6});
 }
 
@@ -416,55 +261,34 @@ TEST_CASE("cartesian_for_index")
 
     cartesian_for_index(f, 2, 3);
     CHECK(res == Vector<int>{0, 0, 0, 0, 1, 2});
+    // Array<int, -2> invalid{};
 }
-
-// todo sequence view
 
 TEST_CASE("head")
 {
-    auto std_array = std::array<int, 3>{1, 2, 3};
-    auto std_vector = std::vector<int>{1, 2, 3};
-    auto array = Array<int, 3>{1, 2, 3};
-    auto arrvec = ArrVec<int, 3>{1, 2, 3};
-    auto vector = Vector<int>{1, 2, 3};
-
-    CHECK(head(std_array) == 1);
-    CHECK(head(std_vector) == 1);
-    CHECK(head(array) == 1);
-    CHECK(head(arrvec) == 1);
-    CHECK(head(vector) == 1);
+    CHECK(head(array_3) == 1);
+    CHECK(head(arrvec_3) == 1);
+    CHECK(head(vector_3) == 1);
 }
 
 TEST_CASE("tail")
 {
     SECTION("const")
     {
-        const auto std_array = std::array<int, 3>{1, 2, 3};
-        const auto std_vector = std::vector<int>{1, 2, 3};
-        const auto array = Array<int, 3>{1, 2, 3};
-        const auto arrvec = ArrVec<int, 3>{1, 2, 3};
-        const auto vector = Vector<int>{1, 2, 3};
-
-        CHECK(tail(std_array) == VectorView<const int>{p_data(std_array) + 1, length(std_array) - 1});
-        CHECK(tail(std_vector) == VectorView<const int>{p_data(std_vector) + 1, length(std_vector) - 1});
-        CHECK(tail(array) == VectorView<const int>{p_data(array) + 1, length(array) - 1});
-        CHECK(tail(arrvec) == VectorView<const int>{p_data(arrvec) + 1, length(arrvec) - 1});
-        CHECK(tail(vector) == VectorView<const int>{p_data(vector) + 1, length(vector) - 1});
+        CHECK(tail(array_3) == ArrayView<const double, 2>{p_data(array_3) + 1});
+        CHECK(tail(arrvec_3) == ArrVecView<const double, 2>{p_data(arrvec_3) + 1, 2});
+        CHECK(tail(vector_3) == VectorView<const double>{p_data(vector_3) + 1, 2, 2});
     }
 
     SECTION("non const")
     {
-        std::array<int, 3> std_array{1, 2, 3};
-        std::vector<int> std_vector{1, 2, 3};
-        Array<int, 3> array{1, 2, 3};
-        ArrVec<int, 3> arrvec{1, 2, 3};
-        Vector<int> vector{1, 2, 3};
+        Array<double, 3> array{1., 2., 3.};
+        ArrVec<double, 3> arrvec{1., 2., 3.};
+        Vector<double> vector{1., 2., 3.};
 
-        CHECK(tail(std_array) == VectorView<int>{p_data(std_array) + 1, length(std_array) - 1});
-        CHECK(tail(std_vector) == VectorView<int>{p_data(std_vector) + 1, length(std_vector) - 1});
-        CHECK(tail(array) == VectorView<int>{p_data(array) + 1, length(array) - 1});
-        CHECK(tail(arrvec) == VectorView<int>{p_data(arrvec) + 1, length(arrvec) - 1});
-        CHECK(tail(vector) == VectorView<int>{p_data(vector) + 1, length(vector) - 1});
+        CHECK(tail(array) == ArrayView<double, 2>{p_data(array) + 1});
+        CHECK(tail(arrvec) == ArrVecView<double, 2>{p_data(arrvec) + 1, 2});
+        CHECK(tail(vector) == VectorView<double>{p_data(vector) + 1, 2, 2});
     }
 }
 
@@ -472,72 +296,52 @@ TEST_CASE("init")
 {
     SECTION("const")
     {
-        const auto std_array = std::array<int, 3>{1, 2, 3};
-        const auto std_vector = std::vector<int>{1, 2, 3};
-        const auto array = Array<int, 3>{1, 2, 3};
-        const auto arrvec = ArrVec<int, 3>{1, 2, 3};
-        const auto vector = Vector<int>{1, 2, 3};
-
-        CHECK(init(std_array) == VectorView<const int>{p_data(std_array), length(std_array) - 1});
-        CHECK(init(std_vector) == VectorView<const int>{p_data(std_vector), length(std_vector) - 1});
-        CHECK(init(array) == VectorView<const int>{p_data(array), length(array) - 1});
-        CHECK(init(arrvec) == VectorView<const int>{p_data(arrvec), length(arrvec) - 1});
-        CHECK(init(vector) == VectorView<const int>{p_data(vector), length(vector) - 1});
+        CHECK(init(array_3) == ArrayView<const double, 2>{p_data(array_3)});
+        CHECK(init(arrvec_3) == ArrVecView<const double, 2>{p_data(arrvec_3), 2});
+        CHECK(init(vector_3) == VectorView<const double>{p_data(vector_3), 2, 2});
     }
 
     SECTION("non const")
     {
-        std::array<int, 3> std_array{1, 2, 3};
-        std::vector<int> std_vector{1, 2, 3};
-        Array<int, 3> array{1, 2, 3};
-        ArrVec<int, 3> arrvec{1, 2, 3};
-        Vector<int> vector{1, 2, 3};
+        Array<double, 3> array{1., 2., 3.};
+        ArrVec<double, 3> arrvec{1., 2., 3.};
+        Vector<double> vector{1., 2., 3.};
 
-        CHECK(init(std_array) == VectorView<int>{p_data(std_array), length(std_array) - 1});
-        CHECK(init(std_vector) == VectorView<int>{p_data(std_vector), length(std_vector) - 1});
-        CHECK(init(array) == VectorView<int>{p_data(array), length(array) - 1});
-        CHECK(init(arrvec) == VectorView<int>{p_data(arrvec), length(arrvec) - 1});
-        CHECK(init(vector) == VectorView<int>{p_data(vector), length(vector) - 1});
+        CHECK(init(array) == ArrayView<double, 2>{p_data(array)});
+        CHECK(init(arrvec) == ArrVecView<double, 2>{p_data(arrvec), 2});
+        CHECK(init(vector) == VectorView<double>{p_data(vector), 2, 2});
     }
 }
 
 TEST_CASE("last")
 {
-    auto std_array = std::array<int, 3>{1, 2, 3};
-    auto std_vector = std::vector<int>{1, 2, 3};
-    auto array = Array<int, 3>{1, 2, 3};
-    auto arrvec = ArrVec<int, 3>{1, 2, 3};
-    auto vector = Vector<int>{1, 2, 3};
-
-    CHECK(last(std_array) == 3);
-    CHECK(last(std_vector) == 3);
-    CHECK(last(array) == 3);
-    CHECK(last(arrvec) == 3);
-    CHECK(last(vector) == 3);
+    CHECK(last(array_3) == 3);
+    CHECK(last(arrvec_3) == 3);
+    CHECK(last(vector_3) == 3);
 }
 
 TEST_CASE("take")
 {
-    SECTION("c array0")
+    SECTION("array 0")
     {
-        const auto res = take(2, c_array_3);
-        CHECK(p_data(res) == p_data(c_array_3));
+        const auto res = take(2, array_3);
+        CHECK(p_data(res) == p_data(array_3));
         CHECK(length(res) == 2);
         CHECK(res[0] == 1.);
         CHECK(res[1] == 2.);
     }
 
-    SECTION("c array1")
+    SECTION("array 1")
     {
-        const auto res = take(-1, c_array_3);
-        CHECK(p_data(res) == p_data(c_array_3));
+        const auto res = take(-1, array_3);
+        CHECK(p_data(res) == p_data(array_3));
         CHECK(length(res) == 0);
     }
 
-    SECTION("c array2")
+    SECTION("array 2")
     {
-        const auto res = take(9, c_array_3);
-        CHECK(p_data(res) == p_data(c_array_3));
+        const auto res = take(9, array_3);
+        CHECK(p_data(res) == p_data(array_3));
         CHECK(length(res) == 3);
         CHECK(res[0] == 1.);
         CHECK(res[1] == 2.);
@@ -547,62 +351,65 @@ TEST_CASE("take")
 
 TEST_CASE("drop")
 {
-    SECTION("c array0")
+    SECTION("array 0")
     {
-        const auto res = drop(2, c_array_3);
-        CHECK(p_data(res) == p_data(c_array_3) + 2);
+        const auto res = drop(2, array_3);
+        CHECK(p_data(res) == p_data(array_3) + 2);
         CHECK(length(res) == 1);
         CHECK(res[0] == 3.);
     }
 
-    SECTION("c array1")
+    SECTION("array 1")
     {
-        const auto res = drop(-1, c_array_3);
-        CHECK(p_data(res) == p_data(c_array_3));
+        const auto res = drop(-1, array_3);
+        CHECK(p_data(res) == p_data(array_3));
         CHECK(length(res) == 3);
+        CHECK(res[0] == 1.);
+        CHECK(res[1] == 2.);
+        CHECK(res[2] == 3.);
     }
 
-    SECTION("c array2")
+    SECTION("array 2")
     {
-        const auto res = drop(9, c_array_3);
-        CHECK(p_data(res) == p_end(c_array_3));
+        const auto res = drop(9, array_3);
+        CHECK(p_data(res) == p_data(array_3) + 3);
         CHECK(length(res) == 0);
     }
 }
 
 TEST_CASE("elem")
 {
-    SECTION("c array")
-    {
-        CHECK(elem(2., c_array_3) == true);
-        CHECK(elem(9., c_array_3) == false);
-    }
-
     SECTION("Array")
     {
         CHECK(elem(2., array_3) == true);
         CHECK(elem(9., array_3) == false);
     }
 
-    SECTION("ArrayView")
+    SECTION("ArrVec")
     {
-        CHECK(elem(2., array_view_3) == true);
-        CHECK(elem(9., array_view_3) == false);
+        CHECK(elem(2., arrvec_3) == true);
+        CHECK(elem(9., arrvec_3) == false);
+    }
+
+    SECTION("Vector")
+    {
+        CHECK(elem(2., vector_3) == true);
+        CHECK(elem(9., vector_3) == false);
     }
 }
 
 TEST_CASE("elem_index")
 {
-    SECTION("c array")
-    {
-        CHECK(elem_index(2., c_array_3).value() == 1);
-        CHECK(elem_index(9., c_array_3).is_nothing() == true);
-    }
-
     SECTION("Array")
     {
         CHECK(elem_index(2., array_3).value() == 1);
         CHECK(elem_index(9., array_3).is_nothing() == true);
+    }
+
+    SECTION("ArrVec")
+    {
+        CHECK(elem_index(2., arrvec_3).value() == 1);
+        CHECK(elem_index(9., arrvec_3).is_nothing() == true);
     }
 
     SECTION("ArrayView")
@@ -614,18 +421,18 @@ TEST_CASE("elem_index")
 
 TEST_CASE("elem_indices")
 {
-    SECTION("c array")
-    {
-        double c_array_3_[3] = {1., 2., 2.};
-        CHECK(elem_indices(2., c_array_3_) == ArrVec<int, 3>{1, 2});
-        CHECK(elem_indices(9., c_array_3_) == ArrVec<int, 3>{});
-    }
-
     SECTION("Array")
     {
-        double array_3_[3] = {1., 2., 2.};
+        const Array<double, 3> array_3_ = {1., 2., 2.};
         CHECK(elem_indices(2., array_3_) == ArrVec<int, 3>{1, 2});
         CHECK(elem_indices(9., array_3_) == ArrVec<int, 3>{});
+    }
+
+    SECTION("ArrVec")
+    {
+        const ArrVec<double, 3> arrvec_3_ = {1., 2., 2.};
+        CHECK(elem_indices(2., arrvec_3_) == ArrVec<int, 3>{1, 2});
+        CHECK(elem_indices(9., arrvec_3_) == ArrVec<int, 3>{});
     }
 
     SECTION("Vector")
@@ -644,16 +451,16 @@ TEST_CASE("find")
     auto is_nine = [](double x)
     { return x == 9. ? true : false; };
 
-    SECTION("c array")
-    {
-        CHECK(find(is_two, c_array_3) == true);
-        CHECK(find(is_nine, c_array_3) == false);
-    }
-
     SECTION("Array")
     {
         CHECK(find(is_two, array_3) == true);
         CHECK(find(is_nine, array_3) == false);
+    }
+
+    SECTION("Vector")
+    {
+        CHECK(find(is_two, vector_3) == true);
+        CHECK(find(is_nine, vector_3) == false);
     }
 
     SECTION("ArrayView")
@@ -671,16 +478,16 @@ TEST_CASE("find_index")
     auto is_nine = [](double x)
     { return x == 9. ? true : false; };
 
-    SECTION("c array")
-    {
-        CHECK(find_index(is_two, c_array_3).value() == 1);
-        CHECK(find_index(is_nine, c_array_3).is_nothing() == true);
-    }
-
     SECTION("Array")
     {
         CHECK(find_index(is_two, array_3).value() == 1);
         CHECK(find_index(is_nine, array_3).is_nothing() == true);
+    }
+
+    SECTION("Vector")
+    {
+        CHECK(find_index(is_two, vector_3).value() == 1);
+        CHECK(find_index(is_nine, vector_3).is_nothing() == true);
     }
 
     SECTION("ArrayView")
@@ -698,23 +505,23 @@ TEST_CASE("find_indices")
     auto is_nine = [](double x)
     { return x == 9. ? true : false; };
 
-    SECTION("c array")
-    {
-        double c_array_3_[3] = {1., 2., 2.};
-        CHECK(find_indices(is_two, c_array_3_) == ArrVec<int, 3>{1, 2});
-        CHECK(find_indices(is_nine, c_array_3_) == ArrVec<int, 3>{});
-    }
-
     SECTION("Array")
     {
-        double array_3_[3] = {1., 2., 2.};
+        const Array<double, 3> array_3_ = {1., 2., 2.};
         CHECK(find_indices(is_two, array_3_) == ArrVec<int, 3>{1, 2});
         CHECK(find_indices(is_nine, array_3_) == ArrVec<int, 3>{});
     }
 
+    SECTION("ArrVec")
+    {
+        const ArrVec<double, 3> arrvec_3_{1., 2., 2.};
+        CHECK(find_indices(is_two, arrvec_3_) == ArrVec<int, 3>{1, 2});
+        CHECK(find_indices(is_nine, arrvec_3_) == ArrVec<int, 3>{});
+    }
+
     SECTION("Vector")
     {
-        Vector<double> vector_3_ = {1, 2, 2};
+        const Vector<double> vector_3_{1., 2., 2.};
         CHECK(find_indices(is_two, vector_3_) == Vector<int>{1, 2});
         CHECK(find_indices(is_nine, vector_3_) == Vector<int>{});
     }
