+++ test/test.cpp
@@ -1,11 +1,9 @@
 #include "sfinae_test.hpp"
-// #include "sequence_test.hpp"
-// #include "prelude_test.hpp"
+#include "sequence_test.hpp"
+#include "prelude_test.hpp"
 // #include "numeric_test.hpp"
 // #include "cyclic_test.hpp"
 // #include "c_utility_test.hpp"
 // #include "enum_type_test.hpp"
 // #include "maybe_test.hpp"
 // #include "scientific_test.hpp"
-#include "crtp_sequence_test.hpp"
-#include "crtp_prelude_test.hpp"
